pyproject.toml
[tool.poetry]
name = "pdf-report-builder"
version = "0.3.6"
description = "Windows app to manage reports structure"
authors = ["karav-n <karav-n@mail.ru>"]
readme = "README.md"
packages = [{include = "pdf_report_builder"}]

[tool.poetry.dependencies]
python = "^3.10,<3.12"
pypdf = "^3.5.2"
wxpython = "^4.2.0"
auto-py-to-exe = "^2.33.0"
pyperclip = "^1.8.2"
pymupdf = "^1.22.3"
reportlab = "^4.0.4"
watchdog = "^3.0.0"
nuitka = "^1.7.3"
python-docx = "^0.8.11"
docx2pdf = "^0.1.8"

[tool.poetry.scripts]
main = "pdf_report_builder.main.py"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
